requirements.txt
streamlit==1.32.0
google-cloud-bigquery==3.12.0
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.36.0
pandas==2.1.0
google-auth>=2.17.0